Update apple_gpu_runtime_stub.cpp
CodeQL #7 fixed. The stub recompiles (with __APPLE__ undefined) and the compute tests pass (4 passed) — the (double) cast is semantics-preserving (it computes the product in double precision, equal-or-better than the old float multiply).

The fix: cv += (double)ckvb[…] * wuvb[…] — casting the first operand to double promotes the whole multiplication to double, so ckvb[…] * wuvb[…] no longer rounds/overflows in float before being accumulated into the double cv. This is the value-multiplication analogue of the earlier index fix (#5 cast j*Dl to size_t; #7 casts the float product to double).

Note: I confirmed the sibling line 1282 (qabs[l] * ckvb[…]) is already safe — qabs is std::vector<double>, so that multiply is already double×float. Only line 1296 had the float×float→double pattern, so this is the complete fix for that alert, not a partial one.

+++ src/compiler/codegen/Tessera_Apple_Backend/runtime/apple_gpu_runtime_stub.cpp
@@ -1293,7 +1293,7 @@ extern "C" void tessera_apple_gpu_mla_absorb_decode_f32(
           double acc = 0;
           for (int32_t j = 0; j < Skv; ++j) {
             double cv = 0;
-            for (int32_t l = 0; l < Dl; ++l) cv += ckvb[(std::size_t)j * Dl + l] * wuvb[(std::size_t)l * dv + d];
+            for (int32_t l = 0; l < Dl; ++l) cv += (double)ckvb[(std::size_t)j * Dl + l] * wuvb[(std::size_t)l * dv + d];
             acc += (score[j] / den) * cv;
           }
           o[d] = (float)acc;
